Update score comparison code to be able to read from and write to our AWS bucket remote storage.
