+++ scripts/utilities/validation/compare_o3_MT_v1.2020_ejam.json
@@ -1,0 +1,13 @@
+{
+  "indicator": "o3",
+  "state": "MT",
+  "version_a": "1.2020",
+  "file_a": "{indicator}/v{version_a}/score_output/final_bg_scores_{state}.csv",
+  "id_a": "block_group_geoid",
+  "score_a": "o3_score",
+  "file_b": "compare/ejamOG/{state}/ejam_{indicator}_subset.csv",
+  "id_b": "ejam_uniq_id",
+  "score_b": "o3",
+  "version_b": "ejam",
+  "out_dir": "compare/output/{state}/{indicator}_v{version_a}_vs_ejam/"
+}
