+++ scripts/utilities/validation/validation_paths.py
@@ -13,6 +13,8 @@ import json
 import os
 import io
 import pandas as pd
+import tempfile
+import importlib
 
 try:
     import boto3
@@ -118,3 +120,85 @@ def read_csv_s3_or_local(path: str, **pd_kwargs) -> pd.DataFrame:
 
     # local file: delegate to pandas
     return pd.read_csv(path, **pd_kwargs)
+
+
+def ensure_local_parent_dir(path: str) -> None:
+    if is_s3_uri(path):
+        return
+    Path(path).parent.mkdir(parents=True, exist_ok=True)
+
+
+def load_fsspec_module():
+    return importlib.import_module('fsspec')
+
+
+def write_df_s3_or_local(df: pd.DataFrame, out_path: str) -> None:
+    ensure_local_parent_dir(out_path)
+    if is_s3_uri(out_path):
+        if boto3 is None:
+            raise RuntimeError("boto3 required to write S3 objects but is not installed")
+        tmp = tempfile.NamedTemporaryFile(delete=False, suffix='.csv')
+        tmp.close()
+        try:
+            df.to_csv(tmp.name, index=False)
+            tail = out_path[5:]
+            parts = tail.split('/', 1)
+            if len(parts) != 2:
+                raise ValueError(f"Invalid S3 URI: {out_path}")
+            bucket, key = parts[0], parts[1]
+            s3 = boto3.client('s3')
+            s3.upload_file(Filename=tmp.name, Bucket=bucket, Key=key)
+        finally:
+            try:
+                os.unlink(tmp.name)
+            except Exception:
+                pass
+        return
+
+    df.to_csv(out_path, index=False)
+
+
+def write_text_s3_or_local(out_path: str, text: str) -> None:
+    ensure_local_parent_dir(out_path)
+    if is_s3_uri(out_path):
+        if boto3 is None:
+            raise RuntimeError("boto3 required to write S3 objects but is not installed")
+        tail = out_path[5:]
+        parts = tail.split('/', 1)
+        if len(parts) != 2:
+            raise ValueError(f"Invalid S3 URI: {out_path}")
+        bucket, key = parts[0], parts[1]
+        s3 = boto3.client('s3')
+        s3.put_object(Bucket=bucket, Key=key, Body=text.encode('utf-8'))
+        return
+
+    with open(out_path, 'w', encoding='utf-8') as fh:
+        fh.write(text)
+
+
+def write_figure_s3_or_local(fig, out_path: str, dpi: int = 150, bbox_inches='tight') -> None:
+    # fig is a matplotlib.figure.Figure
+    if is_s3_uri(out_path):
+        if boto3 is None:
+            raise RuntimeError("boto3 required to write S3 objects but is not installed")
+        tmp = tempfile.NamedTemporaryFile(delete=False, suffix='.png')
+        tmp.close()
+        try:
+            fig.savefig(tmp.name, dpi=dpi, bbox_inches=bbox_inches)
+            tail = out_path[5:]
+            parts = tail.split('/', 1)
+            if len(parts) != 2:
+                raise ValueError(f"Invalid S3 URI: {out_path}")
+            bucket, key = parts[0], parts[1]
+            s3 = boto3.client('s3')
+            s3.upload_file(Filename=tmp.name, Bucket=bucket, Key=key)
+        finally:
+            try:
+                os.unlink(tmp.name)
+            except Exception:
+                pass
+        return
+
+    # local path
+    ensure_local_parent_dir(out_path)
+    fig.savefig(out_path, dpi=dpi, bbox_inches=bbox_inches)
